+++ src/client/components/AuthComponents.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import React, { useState } from "react";
-import { Box, Button, Flex, Heading, jsx, ThemeUIStyleObject } from "theme-ui";
+import { Box, Button, Flex, Heading, Text, jsx, ThemeUIStyleObject } from "theme-ui";
 import { toast } from "react-toastify";
 
 import { Login, JWT, Register } from "../../shared/entities";
@@ -12,26 +12,91 @@ import { userFetch } from "../actions/user";
 import { assertNever } from "../functions";
 import { IProject } from "../../shared/entities";
 import store from "../store";
+import { ReactComponent as SignupIconAudience } from "../media/signup-icon-audience.svg";
+import { ReactComponent as SignupIconDollar } from "../media/signup-icon-dollar.svg";
+import { ReactComponent as SignupIconSearch } from "../media/signup-icon-search.svg";
 
 const style: ThemeUIStyleObject = {
   footer: {
     flex: "auto",
+    display: "flex",
+    flexDirection: "column",
     textAlign: "center",
     fontVariant: "tabular-nums",
     py: 2,
+    mt: 2
+  },
+  footerMicrocopy: {
+    textAlign: "center",
     mt: 2,
     fontSize: 1
   },
-  footerButton: {
-    width: "100%"
-  },
   header: {
-    padding: "16px 12px",
-    margin: "-12px -12px 24px"
+    mb: 7,
+    color: "gray.8"
+  },
+  headerFancy: {
+    position: "relative",
+    zIndex: "0",
+    "&:before": {
+      content: "''",
+      position: "absolute",
+      left: "-35px",
+      right: "-35px",
+      top: "-55px",
+      bottom: "-30px",
+      bg: "warning",
+      opacity: "0.1",
+      zIndex: "-1",
+      transform: "rotate(1deg)"
+    },
+    "&:after": {
+      content: "''",
+      position: "absolute",
+      left: "-35px",
+      right: "-35px",
+      top: "-55px",
+      bottom: "-30px",
+      bg: "warning",
+      opacity: "0.1",
+      zIndex: "-1",
+      transform: "rotate(5deg)"
+    }
+  },
+  heading: {
+    fontFamily: "heading",
+    fontWeight: "light",
+    fontSize: 4
   },
   link: {
     cursor: "pointer",
     color: "blue.5"
+  },
+  featuresList: {
+    listStyleType: "none",
+    pl: 0
+  },
+  featuresListItem: {
+    pl: 0,
+    pb: 4,
+    display: "flex",
+    alignItems: "flex-start",
+    justifyContent: "flex-start",
+    "&:last-of-type": {
+      pb: 0
+    }
+  },
+  featuresListHeading: {
+    fontSize: 3,
+    my: 0
+  },
+  featuresListIcon: {
+    height: "auto",
+    flex: "0 0 40px",
+    mr: 3
+  },
+  featuresListDesc: {
+    my: 2
   }
 };
 
@@ -90,7 +155,9 @@ export const LoginContent = ({ children }: { readonly children: React.ReactNode 
           }}
         />
       </Box>
-      <Button type="submit">Log in</Button>
+      <Button id="primary-action" type="submit">
+        Log in
+      </Button>
     </Flex>
   );
 };
@@ -164,6 +231,7 @@ export const RegisterContent = ({ children }: { readonly children: React.ReactNo
         />
       </Box>
       <Button
+        id="primary-action"
         type="submit"
         disabled={
           ("isPending" in registrationResource && registrationResource.isPending) ||
@@ -183,70 +251,97 @@ export const AuthModalContent = ({ project }: { readonly project: IProject }) =>
 
   const initialContent = (
     <React.Fragment>
-      <Box sx={style.header}>
-        <Heading
-          as="h3"
-          sx={{ marginBottom: "0", fontWeight: "medium", display: "flex", alignItems: "center" }}
-          id="modal-header"
-        >
+      <Box sx={{ ...style.header, ...style.headerFancy }}>
+        <Heading as="h1" sx={style.heading} id="modal-header">
           {project.user.name} builds maps with DistrictBuilder
         </Heading>
+        <Text>
+          DistrictBuilder is a <strong>free</strong> and <strong>open</strong> redistricting tool
+          that puts the power of drawing electoral maps in your hands.
+        </Text>
       </Box>
       <Box>
-        <p>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
-          ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation
-          ullamco laboris nisi ut aliquip ex ea commodo consequat.
-        </p>
-        <ul>
-          <li>100% free to use, forever</li>
-          <li>Makes district drawing easy</li>
-          <li>Another selling point to convince user that this is good</li>
+        <ul sx={style.featuresList}>
+          <li sx={style.featuresListItem}>
+            <SignupIconDollar sx={style.featuresListIcon} />
+            <div>
+              <Heading as="h2" sx={style.featuresListHeading}>
+                100% free, forever
+              </Heading>
+              <Text sx={style.featuresListDesc}>
+                DistrictBuilder is completely free to use and open source.
+              </Text>
+            </div>
+          </li>
+          <li sx={style.featuresListItem}>
+            <SignupIconSearch sx={style.featuresListIcon} />
+            <div>
+              <Heading as="h2" sx={style.featuresListHeading}>
+                Uncover inequalities
+              </Heading>
+              <Text sx={style.featuresListDesc}>
+                Built-in population demographics and real-time updates make drawing fair districts
+                easy.
+              </Text>
+            </div>
+          </li>
+          <li sx={style.featuresListItem}>
+            <SignupIconAudience sx={style.featuresListIcon} />
+            <div>
+              <Heading as="h2" sx={style.featuresListHeading}>
+                Engage your audience, friends, and family
+              </Heading>
+              <Text sx={style.featuresListDesc}>
+                Easily export your custom map. Share it in an article, on social media, or with a
+                friend!
+              </Text>
+            </div>
+          </li>
         </ul>
       </Box>
       <Flex sx={style.footer}>
-        <Button sx={style.footerButton} onClick={() => setModalIntent("register")}>
+        <Button id="primary-action" onClick={() => setModalIntent("register")}>
           Create a free account
         </Button>
+        <Text sx={style.footerMicrocopy}>
+          Already have an account?{" "}
+          <span sx={style.link} onClick={() => setModalIntent("login")}>
+            Log in
+          </span>
+        </Text>
       </Flex>
-      <p sx={style.footer}>
-        Already have an account?{" "}
-        <span sx={style.link} onClick={() => setModalIntent("login")}>
-          Log in
-        </span>
-      </p>
     </React.Fragment>
   );
 
   const loginContent = (
     <React.Fragment>
       <LoginContent>
-        <Heading id="modal-header" as="h2" sx={{ mb: 5, textAlign: "left" }}>
+        <Heading as="h1" sx={{ ...{ mb: 4 }, ...style.heading }} id="modal-header">
           Log in
         </Heading>
       </LoginContent>
-      <Box sx={{ fontSize: 1, textAlign: "center" }}>
+      <Text sx={style.footerMicrocopy}>
         Need an account?{" "}
-        <span sx={style.link} onClick={() => setModalIntent("register")}>
+        <span id="primary-action" sx={style.link} onClick={() => setModalIntent("register")}>
           Sign up for free
         </span>
-      </Box>
+      </Text>
     </React.Fragment>
   );
 
   const registerContent = (
     <React.Fragment>
       <RegisterContent>
-        <Heading id="modal-header" as="h2" sx={{ mb: 5, textAlign: "left" }}>
+        <Heading id="modal-header" as="h1" sx={{ ...{ mb: 4 }, ...style.heading }}>
           Create an account
         </Heading>
       </RegisterContent>
-      <Box sx={{ fontSize: 1, textAlign: "center" }}>
+      <Text sx={style.footerMicrocopy}>
         Already have an account?{" "}
-        <span sx={style.link} onClick={() => setModalIntent("login")}>
+        <span id="primary-action" sx={style.link} onClick={() => setModalIntent("login")}>
           Log in
         </span>
-      </Box>
+      </Text>
     </React.Fragment>
   );